pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain"
version = "1.0.0"
description = "Local-first personal agent harness for an Obsidian vault"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic>=0.25.0",
  "fastapi>=0.110.0",
  "feedparser>=6.0.0",
  "google-api-python-client>=2.120.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "uvicorn>=0.29.0",
  "websockets>=12.0",
]

[project.optional-dependencies]
test = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
]

[project.scripts]
brain = "brain.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["brain*"]
